fix update ship's ticket type and bonuses

+++ src/main/java/ua/bolshak/controller/commands/DeleteRoleCommand.java
@@ -2,6 +2,7 @@ package ua.bolshak.controller.commands;
 
 import ua.bolshak.model.entity.Role;
 import ua.bolshak.model.service.RoleService;
+import ua.bolshak.model.service.UserService;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -12,7 +13,11 @@ public class DeleteRoleCommand implements ICommand {
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         Role role = RoleService.findById(Integer.parseInt(request.getParameter("idRole")));
-        RoleService.delete(role);
+        if (UserService.findAllByRole(role).isEmpty()) {
+            RoleService.delete(role);
+        } else {
+            request.setAttribute("ErrorMassage", "Yhe role has users and can not be deleted!");
+        }
         return new ToUserPage().execute(request, response);
     }
 }
